[Refactor] api 폴더구조 swagger와 동일하게 변경하고 api 경로 수정
