+++ src/app/mydashboard/components/Pagination.tsx
@@ -1,5 +1,5 @@
 import toast from "react-hot-toast";
-import { BackForwardBtn } from "@/components/button/ButtonComponents";
+import { PaginationBtn } from "@/components/button/ButtonComponents";
 import { Dispatch, SetStateAction, useEffect, useState } from "react";
 
 interface IProps {
@@ -9,21 +9,23 @@ interface IProps {
 }
 
 const Pagination = ({ totalPage, setPage, page }: IProps) => {
-  const [isDisabled, setIsDisabled] = useState(false);
+  const [isDisabledNext, setIsDisabledNext] = useState(false);
+  const [isDisabledPrev, setIsDisabledPrev] = useState(false);
 
   useEffect(() => {
-    setIsDisabled(totalPage === page);
+    setIsDisabledNext(totalPage === page);
+    setIsDisabledPrev(totalPage === page);
   }, [page, totalPage]);
 
   const handlePageNext = () => {
     const nextPage = page + 1;
 
     if (nextPage > totalPage) {
-      setIsDisabled(true);
+      setIsDisabledNext(true);
       toast.error("마지막 페이지입니다.");
     } else {
       setPage(nextPage);
-      setIsDisabled(false);
+      setIsDisabledNext(false);
     } 
   }
 
@@ -32,15 +34,15 @@ const Pagination = ({ totalPage, setPage, page }: IProps) => {
 
     if (prevPage < 1) {
       setPage(1);
-      setIsDisabled(true);
+      setIsDisabledPrev(true);
       toast.error("첫 번째 페이지입니다.");
     } else {
       setPage(prevPage);
-      setIsDisabled(false);
+      setIsDisabledPrev(false);
     }
   }
 
-  return <BackForwardBtn onClickNext={handlePageNext} onClickPrev={handlePagePrev} disabled={isDisabled} />
+  return <PaginationBtn onClickNext={handlePageNext} onClickPrev={handlePagePrev} disabledNext={isDisabledNext} disabledPrev={isDisabledPrev}   />
 };
 
 export default Pagination;
